+++ src/Provider/Portugal/Madeira.php
@@ -1,0 +1,41 @@
+<?php
+
+/*
+ * This file is part of the umulmrum/holiday package.
+ *
+ * (c) Stefan Kruppa
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Umulmrum\Holiday\Provider\Portugal;
+
+use Umulmrum\Holiday\Constant\HolidayName;
+use Umulmrum\Holiday\Constant\HolidayType;
+use Umulmrum\Holiday\Model\Holiday;
+use Umulmrum\Holiday\Model\HolidayList;
+use Umulmrum\Holiday\Provider\Religion\ChristianHolidaysTrait;
+
+class Madeira extends Portugal
+{
+    use ChristianHolidaysTrait;
+
+    public function calculateHolidaysForYear(int $year): HolidayList
+    {
+        $holidays = parent::calculateHolidaysForYear($year);
+        if ($year >= 1979) {
+            $holidays->add($this->getMadeiraDay($year, HolidayType::DAY_OFF));
+        }
+        if ($year >= 2002) {
+            $holidays->add($this->getSecondChristmasDay($year, HolidayType::OFFICIAL | HolidayType::DAY_OFF));
+        }
+
+        return $holidays;
+    }
+
+    protected function getMadeiraDay(int $year, int $additionalType = HolidayType::OTHER): Holiday
+    {
+        return Holiday::create(HolidayName::MADEIRA_DAY, "{$year}-07-01", HolidayType::OFFICIAL | $additionalType);
+    }
+}