+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Holidays for Portugal
 - .gitattributes file to filter unnecessary files
 
 ## [0.8.0] - 2024-05-26